makeCallGraph. module reorg.
